sleep, nanosleep added
sleep is nonconformant yet.
No signal handling.
There's the big question,
whether anyone relies on the specifications in this case.
I wouldn't. So, possibly it is ok.
But it is a possible source of a nasty bug.
So I'm adding praeprocessor warnings.
It is again the question of the target of minilib, as well.
As stated somewhere else, minilib is not
about "bordercases".
And somehow it seems to me, relying onto a
conformant sleep IS a bordercase.
And, in all sources I've seen yet,
they didn't rely on the specificated sleep.
Might also be cause of non linux unix systems, bsd, andsoon.
I know. Nasty.
I didn't write the specifications.