Got sed to compile.
It's a milestone.
To me, the size of 18kB of the statically linked binary
is hard to believe - especially, cause the sed
binary, linked dynamically to musl, here
at gentoo comes with 180kB.
But - seems to me, finally all the work pays out.
I have to tidy up, especially the compat headers are quite messed,
and there is some trouble with static/nonstatic predefines.

When I cleaned the sources, I'm going to do a new release.