Fixed a bug dependent on compiler optimization
Syscalls with more than 3 args need to be compiled with
-O0. Otherwise the registers are likely to be optimized
out. Didn't find another solution than passing
-O0 explicitely to the according functions.
Would call this a bug of gcc. But it already is really lacking,
to not have the according asm constraints for x64.
albite x64 is there for. don't know. 20 years.
Anyways. It's fixed for now.