Added the option globals_on_stack
It's possible now to ommit usage of .data, .bss sections,
even for globals. The pointer to the stack is stored within a
global register (r15).
Would be even nicer to have a fixed memory region for the globals,
but since the kernel (linux,amd64) randomizes the stack location,
that's not possible.