Added stack note (remove exec flag from stack)
Had to add the stack program header again.
The stack is loaded an initialized otherwise,
but marked executable.
I personally do believe, there's no risk here -
as long as the whole program is save.
Meaning, when no unsecure functions like scanf
are used, and input is checked, the stack cannot be overwritten.
Anyways.
The option is there. As well as exploits.
So, I added the phdr.

It feels however like having a security hole.
And, instead of fixing the source - things like scanf, gets, andsoon -
doing a quick fix for the known exploit.
What, well, gives the next hole, so let's randomize
the stack location, then put canaries on the stack, andsoon.