Added qsort_pp, sort pointers to pointers.
Did have quite some fun.
Finally got this working.
>; sometimes.. so much work,
in so few lines.
But needed it for my version of scandir,
where I'm trying to omit all unneccessary copying.
And the final few lines of qsort_pp -
I like them. Aesthetical.
For swapping the elements, theres an asm statement.
(xor swap).
The code gcc generated wasn't that nice.

And finally, 1. it is the inner loop.
2.The asm statement is platform portable.
At least afaik, most platform do have a xor
instruction.

If not - it's a simple swap.
And the xor's do two things,
1. one register can be spared, since one operand (source,or dest)
   can stay within memory
2. the temporary var can be omitted.