Extended the regular expression machine.
Now the 'simple' machine suddenly isn't simple anymore.
Going to rename the simple machine to extended,
and redo a simple version.

Overall, I'm satisfied with the regex machine.
It is not the usual logic, but the
thing I wrote has got it's advantages.
Amongst others, the code still is small,
and since there's no backtracking done,
this might outperform the 'regular' regular expression machines.

Besides, this thing is extensible via the $ and % callbacks.