Seems as obviously "opt_fence" is needed for the syscalls.
Added a macro, OPTFENCE(...), to syscalls with 4 or
more parameters.
Before, hard to find bugs by the optimization of gcc occured.,
sometimes.
Possibly the same fence is needed for syscalls < 4 parameters,
as well. Hard to say. Doesn't seem so,
what has to do with how the parameters have to be declared.
(Still no extended asm constraints for registers r10, r8, ..
15 years since x64 got common.) :/

Ok. It is working now.
So I better stop complaining..